fix: isSameSignature on generic method
